More LOD changes on the GUI
Former-commit-id: addf150092d493fdecccdc924620c1b7398515a8

+++ src/apps/conductor/UserPreferences.h
@@ -161,7 +161,7 @@ private:
     std::map< std::string, double > backgroundColor; ///Map the colors to their values
 
     std::map< std::string, bool > preferenceMap; ///<Map to hold preference bools and key names
-    double m_lodScale; ///< The global Geometry LOD scale
+    long m_lodScale; ///< The global Geometry LOD scale
 
     std::string ConvertUnicode( const wxChar* data )
     {
